+++ app/src/main/java/eu/faircode/email/FragmentAccount.java
@@ -122,11 +122,13 @@ public class FragmentAccount extends FragmentBase {
     private CheckBox cbBrowse;
     private CheckBox cbAutoSeen;
     private EditText etInterval;
+    private CheckBox cbNoop;
     private CheckBox cbPartialFetch;
     private CheckBox cbIgnoreSize;
     private RadioGroup rgDate;
     private CheckBox cbUnicode;
     private CheckBox cbUnmetered;
+    private CheckBox cbVpnOnly;
 
     private Button btnCheck;
     private ContentLoadingProgressBar pbCheck;
@@ -233,11 +235,13 @@ public class FragmentAccount extends FragmentBase {
         cbBrowse = view.findViewById(R.id.cbBrowse);
         cbAutoSeen = view.findViewById(R.id.cbAutoSeen);
         etInterval = view.findViewById(R.id.etInterval);
+        cbNoop = view.findViewById(R.id.cbNoop);
         cbPartialFetch = view.findViewById(R.id.cbPartialFetch);
         cbIgnoreSize = view.findViewById(R.id.cbIgnoreSize);
         rgDate = view.findViewById(R.id.rgDate);
         cbUnicode = view.findViewById(R.id.cbUnicode);
         cbUnmetered = view.findViewById(R.id.cbUnmeteredOnly);
+        cbVpnOnly = view.findViewById(R.id.cbVpnOnly);
 
         btnCheck = view.findViewById(R.id.btnCheck);
         pbCheck = view.findViewById(R.id.pbCheck);
@@ -313,6 +317,7 @@ public class FragmentAccount extends FragmentBase {
 
                 etName.setText(position > 1 ? provider.name : null);
                 etInterval.setText(provider.keepalive > 0 ? Integer.toString(provider.keepalive) : null);
+                cbNoop.setChecked(provider.noop);
                 cbPartialFetch.setChecked(provider.partial);
 
                 tvSentWarning.setVisibility(View.GONE);
@@ -899,12 +904,14 @@ public class FragmentAccount extends FragmentBase {
         args.putBoolean("browse", cbBrowse.isChecked());
         args.putBoolean("auto_seen", cbAutoSeen.isChecked());
         args.putString("interval", etInterval.getText().toString());
+        args.putBoolean("noop", cbNoop.isChecked());
         args.putBoolean("partial_fetch", cbPartialFetch.isChecked());
         args.putBoolean("ignore_size", cbIgnoreSize.isChecked());
         args.putBoolean("use_date", rgDate.getCheckedRadioButtonId() == R.id.radio_date_header);
         args.putBoolean("use_received", rgDate.getCheckedRadioButtonId() == R.id.radio_received_header);
         args.putBoolean("unicode", cbUnicode.isChecked());
         args.putBoolean("unmetered", cbUnmetered.isChecked());
+        args.putBoolean("vpn_only", cbVpnOnly.isChecked());
 
         args.putSerializable("drafts", drafts);
         args.putSerializable("sent", sent);
@@ -971,12 +978,14 @@ public class FragmentAccount extends FragmentBase {
                 boolean browse = args.getBoolean("browse");
                 boolean auto_seen = args.getBoolean("auto_seen");
                 String interval = args.getString("interval");
+                boolean noop = args.getBoolean("noop");
                 boolean partial_fetch = args.getBoolean("partial_fetch");
                 boolean ignore_size = args.getBoolean("ignore_size");
                 boolean use_date = args.getBoolean("use_date");
                 boolean use_received = args.getBoolean("use_received");
                 boolean unicode = args.getBoolean("unicode");
                 boolean unmetered = args.getBoolean("unmetered");
+                boolean vpn_only = args.getBoolean("vpn_only");
 
                 EntityFolder drafts = (EntityFolder) args.getSerializable("drafts");
                 EntityFolder sent = (EntityFolder) args.getSerializable("sent");
@@ -1075,6 +1084,8 @@ public class FragmentAccount extends FragmentBase {
                         return true;
                     if (!Objects.equals(account.poll_interval, poll_interval))
                         return true;
+                    if (!Objects.equals(account.keep_alive_noop, noop))
+                        return true;
                     if (!Objects.equals(account.partial_fetch, partial_fetch))
                         return true;
                     if (!Objects.equals(account.ignore_size, ignore_size))
@@ -1086,6 +1097,8 @@ public class FragmentAccount extends FragmentBase {
                     if (!Objects.equals(account.unicode, unicode))
                         return true;
                     if (unmetered != jconditions.optBoolean("unmetered"))
+                        return true;
+                    if (vpn_only != jconditions.optBoolean("vpn_only"))
                         return true;
                     if (account.error != null && account.synchronize)
                         return true;
@@ -1222,7 +1235,7 @@ public class FragmentAccount extends FragmentBase {
                         account.keep_alive_succeeded = 0;
                     }
                     account.poll_interval = Math.max(1, poll_interval);
-
+                    account.keep_alive_noop = noop;
                     account.partial_fetch = partial_fetch;
                     account.ignore_size = ignore_size;
                     account.use_date = use_date;
@@ -1231,6 +1244,7 @@ public class FragmentAccount extends FragmentBase {
                     account.unicode = unicode;
 
                     jconditions.put("unmetered", unmetered);
+                    jconditions.put("vpn_only", vpn_only);
                     account.conditions = jconditions.toString();
 
                     if (!update)
@@ -1588,10 +1602,12 @@ public class FragmentAccount extends FragmentBase {
                     cbBrowse.setChecked(account == null ? true : account.browse);
                     cbAutoSeen.setChecked(account == null ? true : account.auto_seen);
                     etInterval.setText(account == null ? "" : Long.toString(account.poll_interval));
+                    cbNoop.setChecked(account == null ? true : account.keep_alive_noop);
                     cbPartialFetch.setChecked(account == null ? true : account.partial_fetch);
                     cbIgnoreSize.setChecked(account == null ? false : account.ignore_size);
                     cbUnicode.setChecked(account == null ? false : account.unicode);
                     cbUnmetered.setChecked(jcondition.optBoolean("unmetered"));
+                    cbVpnOnly.setChecked(jcondition.optBoolean("vpn_only"));
 
                     if (account != null && account.use_date)
                         rgDate.check(R.id.radio_date_header);